cleanup pr - revert unneeded changes in package.json - revert files moved to src - linting
