+++ index.js
@@ -1,13 +1,13 @@
-import * as poParser from './poparser.js';
-import poCompiler from './pocompiler.js';
-import moParser from './moparser.js';
-import moCompiler from './mocompiler.js';
+import * as poParser from './lib/poparser.js';
+import poCompiler from './lib/pocompiler.js';
+import moParser from './lib/moparser.js';
+import moCompiler from './lib/mocompiler.js';
 
 /**
  * Translation parser and compiler for PO files
  * @see https://www.gnu.org/software/gettext/manual/html_node/PO.html
  *
- * @type {import("./index.js").PoParser} po
+ * @type {import("./index.d.ts").po} po
  */
 export const po = {
   parse: poParser.parse,
@@ -19,9 +19,11 @@ export const po = {
  * Translation parser and compiler for PO files
  * @see https://www.gnu.org/software/gettext/manual/html_node/MO.html
  *
- * @type {import("./index.js").MoParser} mo
+ * @type {import("./index.d.ts").mo} mo
  */
-export const mo= {
+export const mo = {
   parse: moParser,
   compile: moCompiler
 };
+
+export default { mo, po };
