+++ lib/mocompiler.js
@@ -1,13 +1,13 @@
 import { Buffer } from 'safe-buffer';
 import encoding from 'encoding';
-import { compareMsgid, formatCharset, generateHeader, HEADERS } from './shared.js';
+import { HEADERS, formatCharset, generateHeader, compareMsgid } from './shared.js';
 import contentType from 'content-type';
 
 /**
  * Exposes general compiler function. Takes a translation
  * object as a parameter and returns binary MO object
  *
- * @param {import('./index.js').gettextTranslations} table Translation object
+ * @param {import('./index.d.ts').GetTextTranslations} table Translation object
  * @return {Buffer} Compiled binary MO object
  */
 export default function (table) {
@@ -20,10 +20,10 @@ export default function (table) {
  * Creates a MO compiler object.
  *
  * @constructor
- * @param {import('./index.js').gettextTranslations} table Translation table as defined in the README
- * @return {import('./index.js').Compiler} Compiler
+ * @param {import('./index.d.ts').GetTextTranslations} table Translation table as defined in the README
+ * @return {import('./index.d.ts').Compiler} Compiler
  */
-function Compiler (table ) {
+function Compiler (table = {}) {
   this._table = table;
 
   let { headers = {}, translations = {} } = this._table;
@@ -71,8 +71,6 @@ function Compiler (table ) {
   this._writeFunc = 'writeUInt32LE';
 
   this._handleCharset();
-
-  return this;
 }
 
 /**
@@ -151,7 +149,7 @@ Compiler.prototype._generateList = function () {
 /**
  * Calculate buffer size for the final binary object
  *
- * @param {import('./index.js').GettextTranslations} list An array of translation strings from _generateList
+ * @param {import('./index.d.ts').GetTextTranslations} list An array of translation strings from _generateList
  * @return {Object} Size data of {msgid, msgstr, total}
  */
 Compiler.prototype._calculateSize = function (list) {
@@ -186,7 +184,7 @@ Compiler.prototype._calculateSize = function (list) {
 /**
  * Generates the binary MO object from the translation list
  *
- * @param {import('./index.js').GettextTranslations} list translation list
+ * @param {import('./index.d.ts').GetTextTranslations} list translation list
  * @param {Object} size Byte size information
  * @return {Buffer} Compiled MO object
  */
@@ -240,7 +238,7 @@ Compiler.prototype._build = function (list, size) {
 };
 
 /**
- * Compiles the translation object into a binary MO object
+ * Compiles translation object into a binary MO object
  *
  * @return {Buffer} Compiled MO object
  */
