+++ lib/pocompiler.js
@@ -1,6 +1,8 @@
-import { compareMsgid, foldLine, formatCharset, generateHeader, HEADERS } from './shared.js';
-import contentType from 'content-type';
+import { Buffer } from 'safe-buffer';
 import encoding from 'encoding';
+import { HEADERS, foldLine, compareMsgid, formatCharset, generateHeader } from './shared.js';
+import contentType from 'content-type';
+
 
 /**
  * Exposes general compiler function. Takes a translation
@@ -72,7 +74,7 @@ function Compiler (table = {}, options = {}) {
  * @param {Object} comments A comments object
  * @return {String} A comment string for the PO file
  */
-Compiler.prototype._drawComments = function (comments){
+Compiler.prototype._drawComments = function (comments) {
   const lines = [];
   const types = [{
     key: 'translator',
