+++ lib/poparser.js
@@ -68,7 +68,7 @@ Parser.prototype.parse = function () {
 /**
  * Detects charset for PO strings from the header
  *
- * @param buf PO string buffer to be parsed
+ * @param {Buffer} buf headers Header value
  */
 Parser.prototype._handleCharset = function (buf = '') {
   const str = buf.toString();
