+++ package.json
@@ -17,24 +17,15 @@
   "engines": {
     "node": ">=18"
   },
-  "files": [
-    "lib",
-    "test",
-    "*.json",
-    "*.md",
-    "license.txt"
-  ],
   "scripts": {
-    "start": "tsc --watch",
-    "build": "tsc",
-    "lint": "eslint src/*.js test/*.js",
+    "lint": "eslint lib/*.js test/*.js index.js",
     "test-generate-mo": "msgfmt test/fixtures/latin13.po -o test/fixtures/latin13.mo & msgfmt test/fixtures/utf8.po -o test/fixtures/utf8.mo & msgfmt test/fixtures/obsolete.po -o test/fixtures/obsolete.mo",
     "test": "mocha",
     "preversion": "npm run lint && npm test",
-    "postversion": "git push && git push --tags"
+    "postversion": "git push && git push --tags",
+    "prepublish": "tsc && npm run lint && npm run test"
   },
-  "main": "./lib/index.js",
-  "types": "./lib/index.d.ts",
+  "main": "./index.js",
   "license": "MIT",
   "dependencies": {
     "content-type": "^1.0.5",
@@ -46,15 +37,16 @@
     "@types/chai": "latest",
     "@types/content-type": "^1.1.8",
     "@types/mocha": "latest",
+    "@types/readable-stream": "^4.0.11",
     "@typescript-eslint/eslint-plugin": "^6.18.1",
     "@typescript-eslint/parser": "^6.14.0",
-    "chai": "^5.1.0",
+    "chai": "^5.0.3",
     "eslint": "^8.56.0",
     "eslint-config-standard": "^17.1.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-n": "^16.6.2",
     "eslint-plugin-promise": "^6.1.1",
-    "mocha": "^10.4.0",
+    "mocha": "^10.3.0",
     "typescript": "^5.4.4"
   },
   "keywords": [
