+++ tsconfig.json
@@ -2,36 +2,35 @@
   "compilerOptions": {
     "noImplicitAny": true,
     "removeComments": false,
-    "module": "ESNext",
-    "moduleResolution": "Node",
-    "target": "ES2021",
-    "declaration": true,
-    "emitDeclarationOnly": false,
-    "preserveConstEnums": true,
-    "sourceMap": true,
-    "strict": false,
+    "module": "Node16",
+    "moduleResolution": "Node16",
+    "target": "ES2015",
+
+    // Strict mode
+    "strict": true,
+
+    // Allow javascript files
     "allowJs": true,
-    "downlevelIteration": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "allowSyntheticDefaultImports": true,
-    "experimentalDecorators": true,
-    "rootDir": "src",
-    "outDir": "lib",
+
+    // Check js files for errors
+    "checkJs": false,
+
+    // Output d.ts files to @types
+    "outDir": "@types",
+
+    // Generate d.ts files
+    "declaration": true,
+
+    // This compiler run should
+    // only output d.ts files
+    "emitDeclarationOnly": true,
     // Minify
     "pretty": false,
-    // types
-    "typeRoots": [
-      "src/types.d.ts",
-      "node_modules/@types",
-      "@types"
-    ],
-    "lib": [
-      "dom",
-      "ESNext"
-    ]
+    // Skip lib check when compiling
+    "skipLibCheck": true
   },
   "include": [
-    "./src/**/*"
+    "lib/**/*.js",
+    "index.d.ts"
   ]
 }
